// src/app/components/types.ts
export interface userDef {
  user?: string;
  location: string;
  name: string;
  number: string;
  alt_number: string;
  email: string;

  //   for address
  house: string;
  area: string;
  landmark: string;
  pincode: string;
  //
  gender: string;
  age: string;
  illness: string;
  gpsLocation: string;
  adhar: string;
  pan: string;
  transport: string;
}